Change dashboard asset path manipulation
